Replace static_cast<int|i64>(...size()) with ssize() commit_hash:889f1aedbb4575db063289770c7a414cdc8d8fd6

+++ yt/yt/server/node/data_node/p2p.cpp
@@ -361,7 +361,7 @@ std::vector<TP2PSuggestion> TP2PSnooper::OnBlockRead(
             }
         }
 
-        if (blockIsHot && (*blocks)[i].Size() >= static_cast<size_t>(config->MaxBlockSize)) {
+        if (blockIsHot && (*blocks)[i].Size() >= config->MaxBlockSize) {
             ThrottledBytes_.Increment((*blocks)[i].Size());
             ThrottledLargeBlockBytes_.Increment((*blocks)[i].Size());
             (*blocks)[i] = {};
